Update cache and factoring
