+++ scripts/update_cache.py
@@ -1,0 +1,6 @@
+# coding: utf-8
+"""This script update a cache with a list of OPI"""
+import cache
+
+if __name__ == "__main__":
+    cache.generate(update=True)
